pyproject.toml
[tool.poetry]
name = "comp0087"
version = "0.1.0"
description = ""
authors = ["Marcell O <omarcell@yahoo.com>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.10"
pandas = "^1.5.3"
openai = "^0.26.5"
tiktoken = "^0.2.0"
tls-client = "^0.1.8"
undetected-chromedriver = "^3.4.6"
selenium = "^4.8.0"
OpenAIAuth = ">0.1.0"
httpx = "^0.23.3"
openpyxl = "^3.1.1"
transformers = "^4.26.1"
torch = "^1.13.1"
python-dateutil = "^2.8.2"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
